chore: flattened claim feature & merged few configs
