--- config/locale.ts
@@ -1,1 +1,0 @@
-export const LOCALE = 'en-US';
