--- features/claim/claim-form/index.ts
@@ -1,1 +1,0 @@
-export * from './claim-form';
